+++ c#/Compiler/Compiler/obj/Debug/net10.0/Compiler.AssemblyInfo.cs
@@ -14,7 +14,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("Compiler")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+fffedaf79153166e7dc40fd1d4f464bc11d5c5dc")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+86b6780a3566ca40fd8b93ab2209cb9d3647b334")]
 [assembly: System.Reflection.AssemblyProductAttribute("Compiler")]
 [assembly: System.Reflection.AssemblyTitleAttribute("Compiler")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]